Begin amble-lsp refactor -- split monolith into modules
